Change the test circumfix yet again
I should have thought about the last circumfix change before making it.
If someone starts a test name with an underscore, they might
inadvertently invoke UB because the test symbol would have two
underscores in it. This change removes the underscore at the end of the
prefix. Hopefully this will be the last time the circumfix changes.