Use pipes less vulnerable to blocking
Before this, tests that output a very large amount of text would hang
the entire program because the output pipe would buffer while the main
process was waiting for them to finish. This change stops using actual
pipes for this kind of thing. Instead, a temporary file is created. It
is opened for reading and writing (the two ends of the pipe.) The file
is then unlinked. This pipe does not suffer from the buffering problem.